adds pages Booking and Home and reorders file structure to put App in src directory
