+++ client/src/pages/Booking.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Calendar from '../components/Calendar';
+
+const Booking = () => (
+  <div>
+    <Calendar />
+  </div>
+);
+
+export default Booking;
