+++ client/src/pages/Home.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const Home = () => (
+  <div>
+    <Link to="/booking">Book Now</Link>
+  </div>
+);
+
+export default Home;
