+++ package.json
@@ -46,6 +46,7 @@
     "moment": "^2.29.1",
     "pm2": "^5.1.0",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.0"
   }
 }
